saas-rucha: Fix POS UI of mobile and wev, Fix Products and Category issue, Add No data found screen for each feature, Fix get product with category model class .

+++ lib/data/models/POS/cart_product_model.dart
@@ -3,10 +3,10 @@ class CartItemModel {
   final String name;
   final String image;
   final String unit;
-  int cost;
+  double cost;
   int count;
-  final int quantity;
-  int variantCost;
+  final double quantity;
+  double variantCost;
 
   CartItemModel(
       {required this.id,
